fixed import errors, moved around some code
